Improved error handling in post requests, changed content length type

+++ headers/Request.hpp
@@ -37,7 +37,7 @@ class Request
         std::map<std::string, std::string>  _headers;
         std::string                         _body;
         int                                 _port;
-        int                                 _contentLength;
+        unsigned long long                  _contentLength;
 
     public:
         Request() = delete;
@@ -53,7 +53,7 @@ class Request
     std::map<std::string, std::string>  getHeaders() const;
     std::string                         getBody() const;
     int                                 getPort() const;
-    int                                 getContentLength() const;
+    unsigned long long                  getContentLength() const;
 
     void    setMethod(std::string method);
     void    setUri(std::string uri);
@@ -64,7 +64,7 @@ class Request
     void    setHeader(std::string headerName, std::string headerValue);
     void    setBody(std::string body);
     void    setPort(int port);
-    void    setContentLength(int contentLength);
+    void    setContentLength(unsigned long long contentLength);
 
     void    parseURI(std::string uri);
     void    parsePostRequest(std::stringstream& ss);
